==> 1. Data wrangling.ipynb <==
# 4.4 Data Wrangling

# Contents
# 1. Drop column
# 2. Find missing values
# 3. Rename column
# 4. Change data type
# 5. Transpose columns to rows
# 6. Rename index columns
# 7. Create data dictionary
# 8. Create data subset
# 9. 1-4 & 8 applied to orders data
# 10. Export new dataframes

# import libraries
import pandas as pd
import numpy as np
import os

#connect to products data source
path = r'/Users/alison/Documents/Data Analysis - Tasks/Instacart  Basket Analysis' 
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

# check data dimensions
df_prods.info

# connect to orders data source
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), index_col = False)

# check orders data dimensions
df_ords.info

# 1. Drop column

# dropping eval set column from orders.csv
df_ords =df_ords.drop(columns = ['eval_set'])

# 2. Find missing values 

# look for missing values in days since prior orders column
df_ords['days_since_prior_order'].value_counts(dropna = False)

# 3. Rename column

# rename column
df_ords.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

# check changes
df_ords.head()

# 4. Change data type

#check dtypes
df_ords.dtypes

# change order id from int to obj
df_ords['order_id'] = df_ords['order_id'].astype('str')

#check change dtype
df_ords['order_id'].dtype

#connect to departments data source
df_deps = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'departments.csv'), index_col = False)

# check column names
df_deps.head()

# 5. Transpose columns to rows

# transposing columns to rows
df_deps.T

# save new format
df_deps_t = df_deps.T

# 6. Rename index columns

df_deps_t.reset_index()

#1) Create a new header; 2) Remove the first row in the dataframe; and 3) Add the new header.
new_header = df_deps_t.iloc[0]
df_deps_t_new = df_deps_t[1:]
df_deps_t_new.columns = new_header

# check column names
df_deps_t_new.head()

# 7. Create data dictionary

#data dictionary
data_dict = df_deps_t_new.to_dict('index')

data_dict

# Use data dictionary to determine product type
print(data_dict.get('19'))

# 8. Create data subset

# create data subset of snacks
df_snacks =  df_prods[df_prods['department_id']==19]

# check column names
df_snacks.head()

# alternative way to create subset snacks
df_snacks_2 = df_prods.loc[df_prods['department_id'] == 19]

df_snacks_2.head(2)

# create same subset using is in ( good for more than one value)
df_snacks_3 = df_prods.loc[df_prods['department_id'].isin([19])]

df_snacks_3.head()

# 9. Above steps applied to orders dataframe

# Change data type
#2. Find another identifier variable in the df_ords dataframe that doesn’t need to be included in your analysis as a numeric variable and change it to a suitable format.
df_ords['user_id'] = df_ords['user_id'].astype('str')

# check user_id changed from int to object
df_ords.dtypes

# Rename column
#3. Look for a variable in your df_ords dataframe with an unintuitive name and change its name without overwriting the dataframe.
df_ords.rename(columns = {'days_since_prior_order' : 'days_since_last_order'})

# Business Analysis
# 4. Your client wants to know what the busiest hour is for placing orders. 
# Find the frequency of the corresponding variable and share your findings.
df_ords['order_hour_of_day'].value_counts()

# Answer to 4. 
# 10 am is the busiest hour for placing orders.
# Next business question
#5. Determine the meaning behind a value of 4 in the "department_id" column within the df_prods dataframe using a data dictionary.
print(data_dict.get('4'))

# Answer to 5. 
# the value 4 in the department_id column refers to produce (code to build dictionary in pre-task area above)
# Next business question
#6. The sales team in your client’s organization wants to know more about breakfast item sales. Create a subset containing only the required information.
# get numerical code for breakfast items
data_dict

# create breakfast data subset
# taking this task as literal 'breakfast' labelled item and ignoring other categories that could also be breakfast foods.
df_breakfast =  df_prods[df_prods['department_id']==14]

# check breakfast column names
df_breakfast.head()

#Next business question
#7. They’d also like to see details about products that customers might use to throw dinner parties. 
# Your task is to find all observations from the entire dataframe that include items from the following departments: 
# alcohol, deli, beverages, and meat/seafood. You’ll need to present this subset to your client.
df_dinner_parties = df_prods.loc[df_prods['department_id'].isin([5,7,12,20])]

# check dinner parties column names
df_dinner_parties.head()

# check dinner parties data dimensions
#8. It’s important that you keep track of total counts in your dataframes. How many rows does the last dataframe you created have?
df_dinner_parties.shape

# Answer to 8. 
#the dinner parties df has 7650 rows. 
# Next business question
#9. Someone from the data engineers team in Instacart thinks they’ve spotted something strange about the customer with a "user_id" of “1.” 
# Extract all the information you can about this user.
# create data subset for user 1
df_user_1 =  df_ords[df_ords['user_id']=='1']

# check user_1 columns
df_user_1.head()

# next business question
# You also need to provide some details about this user’s behavior. 
# What basic stats can you provide based on the information you have?
df_user_1.info

# conduct value counts on columns
df_user_1['orders_day_of_week'].value_counts()

# Partial answer to 9.
# There are 11 entries for user 1. They place an order most often on a Wednesday (4) then Sunday (3). 
# note DOW code is 0 to 6: Saturday to Friday
df_user_1['order_hour_of_day'].value_counts()

# Partial answer to 9.
# User 1 places orders more frequently at 7 or 8 in the morning, and never after 4pm.
df_user_1['days_since_prior_order'].value_counts()

# Partial answer to 9.
#they place orders irregularly but within a month. They have once placed two orders on the same day.

# 10. Export new dataframes

#12. Export your df_ords dataframe as “orders_wrangled.csv” in your “Prepared Data” folder.
df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_wrangled.csv'))

# Export the df_dep_t_new dataframe as “departments_wrangled.csv” in your “Prepared Data” folder 
# so that you have a “.csv” file of your departments data in the correct format.
df_deps_t_new.to_csv(os.path.join(path, '02 Data','Prepared Data', 'departments_wrangled.csv'))
