==> 2. Consistency checks.ipynb <==
# 4.5 Consistency checks

# Contents
# 1. Find missing values
# 2. Find duplicates
# 3. Remove duplicates
# 4. Repeat 1-3 on the Orders df
# 5. Create new customer flag


#import libraries
import pandas as pd
import numpy as np
import os

path = r'/Users/alison/Documents/Data Analysis - Tasks/Instacart  Basket Analysis'
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

path = r'/Users/alison/Documents/Data Analysis - Tasks/Instacart  Basket Analysis'
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_wrangled.csv'), index_col = 0)

# check shape of ords df
df_ords.shape

df_ords.describe()

# check shape
df_prods.shape

# 1. Find missing values

#find missing values in prods df
df_prods.isnull().sum()

df_nan = df_prods[df_prods['product_name'].isnull() == True]

df_nan

# missing values cannot be imputed and total 16. Remove missing values.
df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

# check missing values were removed. Should be 16 less rows.
df_prods_clean.shape

# 2. Find duplicates

df_dups = df_prods_clean[df_prods_clean.duplicated()]

df_dups

# 3. Remove duplicates

# There are five duplicates which can be removed. Remove duplicates.
df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

# check duplicates removed. Should be five less rows.
df_prods_clean_no_dups.shape

# export cleaned prods df
df_prods_clean_no_dups.to_csv(os.path.join(path, '02 Data','Prepared Data', 'products_checked.csv'))

# 4. Repeat 1-3 on the Orders df

df_ords.describe()

#notes about the df_ords data above: days of the week min 0, max 6 as expected; hour of day min 00:00, max 23:23); days since last order min 0, max 303 days); nothing seems out of order.
# check for mixed  data types
for col in df_ords.columns.tolist():
  weird = (df_ords[[col]].map(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[weird]) > 0:
    print (col)

#no mixed data in df_ords
#5. Run a check for missing values in your df_ords dataframe.
df_ords.isnull().sum()

df_nan2 = df_ords[df_ords['days_since_last_order'].isnull() == True]

df_nan2

# 5. Create new customer flag

# These missing values might be new customer accounts without any existing orders since 0 days would record two orders on the same day. 
# There are 206209 rows in the days since last order column with missing values. Check also that the order number is 1.
# Need to create a new variable that acts like a flag based on the missing value. 
# First create the 'New_customer' column
def New_customer(row):
    if pd.isna(row['days_since_prior_order']) and (row['order_number'] == 1):
        return 'Yes'
    else:
        return 'No'

df_ords['New_customer_flag'] = df_ords.apply(New_customer, axis=1)

print (df_ords)

#check of change made. Yes should be 206,209.
df_ords['New_customer_flag'].value_counts()

# check for duplicates
df_ords_clean = df_ords[df_ords['user_id'].isnull() == False]

df_dups2 = df_ords_clean[df_ords_clean.duplicated()]

df_dups2

# no duplicates found
# confirm data dimensions of new ords df
df_ords_clean.shape

# export clean ords df
df_ords_clean.to_csv(os.path.join(path, '02 Data','Prepared Data', 'ords_checked.csv'))
