==> 3. Combine Dataframes.ipynb <==
# 4.6 Combine dataframes

# Contents
# 1. Merge dataframes 
# 2. Export as pickle file

#import libraries
import pandas as pd
import numpy as np
import os

# define path
path = r'/Users/alison/Documents/Data Analysis - Tasks/Instacart  Basket Analysis'

# get new df
df_ords_prior = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders_products_prior.csv'), index_col = False)

df_ords_prior.head()

# get latest ords df
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'ords_checked.csv'), index_col = 0)

df_ords.head()

# check shape from last export. Expecting (3421083, 7)
df_ords.shape

# check dimensions of df_ords_prior df
df_ords_prior.shape

# 1. Merge dataframes

# merge tables smaller first inner join
df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id', indicator = True)

# check columns on new table
df_merged_large.head()

# check if full match
df_merged_large['_merge'].value_counts()

# 2. Export as pickle file

# Export merged df to pkl
df_merged_large.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_combined.pkl'))
