==> 4. Combine another dataframe.ipynb <==
# Merge with a third table

# Contents
# 1. Merge dataframes
# 2. Export new table as pickle file

#import libraries
import pandas as pd
import numpy as np
import os

# define path to data folders
path = r'/Users/alison/Documents/Data Analysis - Tasks/Instacart  Basket Analysis'

# get previous merged table
df_combined = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_combined.pkl'))

# 4. Check the shape of the imported dataframe. Expecting (32434489, 11)
df_combined.shape

# can confirm shape same as when created
# get latest products df
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'products_checked.csv'), index_col = 0)

# check table dimensions. expecting (49672, 5)
df_prods.shape

# 1. Merge dataframes

# 5. Determine a suitable way to combine the orders_products_combined dataframe with products data set.
# can merge on product_id
# merge tables on product id, smallest first inner join
df_merged_mega = df_prods.merge(df_combined, on = 'product_id')

# 6. Confirm the results of the merge using the merge flag.
df_merged_mega['_merge'].value_counts()

df_merged_mega.head()

# check new table dimensions
df_merged_mega.shape

# 2. Export new df as pickle file

# 7. Export the newly created dataframe as ords_prods_merge
df_merged_mega.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'ords_prods_merge.pkl'))
