==> 5. Derive New Variables.ipynb <==
# Deriving new variables

# Contents
# 1. Create a subset
# 2. Create if, or else statements to group data
# 3. Create new column to retain grouped data
# 4. repeat 2 and 3 for new variable

#import libraries
import pandas as pd
import numpy as np
import os

# define path to data source
path = r'/Users/alison/Documents/Data Analysis - Tasks/Instacart  Basket Analysis'

# get latest merged df
df_ords_prods_merge = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ords_prods_merge.pkl'))

# 1. Create a subset

# create a subset
df = df_ords_prods_merge[:1000000]

# 2. Create if, or else statements to group data

#If the value within the “prices” column within the given row is less than or equal to 5, then return the string “Low-range product.”

#Or else, if the value within the “prices” column within the given row is greater than 5 and less than or equal to 15, then return the string “Mid-range product.”

#Or else, if the value of the “prices” column within the given row is greater than 15, then return the string “High-range product.”

#Or else, return the string “Not enough data.”
df_ords_prods_merge.loc[df_ords_prods_merge['prices'] > 15, 'price_range_loc'] = 'High-range product'

df_ords_prods_merge.loc[(df_ords_prods_merge['prices'] <= 15) & (df_ords_prods_merge['prices'] > 5), 'price_range_loc'] = 'Mid-range product' 

df_ords_prods_merge.loc[df_ords_prods_merge['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df_ords_prods_merge['price_range_loc'].value_counts(dropna=False)

# Group data into two busiest days and two slowest days
df_ords_prods_merge['orders_day_of_week'].value_counts(dropna=False)

# If the value is 0 or 1 then assign Busiest days
# if the value is 3 or 4 then assign Slowest days
result = []

for value in df_ords_prods_merge["orders_day_of_week"]:
    if value in [0, 1]:
        result.append("Busiest days")
    elif value in [3, 4]:
        result.append("Slowest days")
    else:
        result.append("Regularly busy")

result

# 3. Create new column to retain grouped data

#2. Create a new column for this using a suitable method.
df_ords_prods_merge["Busiest_days"] = result

# print frequency of new column
df_ords_prods_merge["Busiest_days"].value_counts(dropna=False)

# 4. repeat 2 and 3 for new variable

# identify the busiest hour of the day
df_ords_prods_merge['order_hour_of_day'].value_counts(dropna=False)

# group busiest period of day into “Most orders,” “Average orders,” and “Fewest orders.”
# busiest period of day: if >2,000,000 09:00 to 17:00 then 'most orders'  if <500,000 23:00 to 6:00 'fewest orders' else 'average orders'
result = []

for value in df_ords_prods_merge['order_hour_of_day']:
    if value in [10, 11, 14, 15, 13, 12, 16, 9]:
        result.append('most_orders')
    elif value in [23, 0, 1, 2, 3, 4, 5, 6]:
        result.append('fewest orders')
    else:
        result.append('average orders')

#create new column to group busiest period of day
df_ords_prods_merge['busiest period of day'] = result

# print frequency of new column
df_ords_prods_merge['busiest period of day'].value_counts(dropna=False)

# check new df dimensions
df_ords_prods_merge.shape

# Export data to pkl
df_ords_prods_merge.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'new_ords_prods_merge.pkl'))
