==> 6. Group data and aggregate variables.ipynb <==
# Grouping data and aggregating variables

# Contents
# 1. Use group by and agg to group data
# 2. Create new categories
# 3. Create flag

#import libraries
import pandas as pd
import numpy as np
import os
# def file path
path = r'/Users/alison/Documents/Data Analysis - Tasks/Instacart  Basket Analysis' 

#get df ords_prods_merged
ords_prods_merge = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'new_ords_prods_merge.pkl'))

df = ords_prods_merge[:1000000]

df.shape

# check column names
df.head()

# 1. Use group by and agg to group data

df.groupby('product_name')

df.groupby('department_id').agg({'order_number': ['mean']})

df.groupby('department_id')['order_number'].mean()

# get min, max, mean
df.groupby('department_id').agg({'order_number': ['mean', 'min', 'max']})

# 2. Create new categories

# To create your flag, you’ll need some criteria. You can use the following:
# If the maximum orders the user has made is over 40, then the customer will be labeled a “Loyal customer.”
# If the maximum orders the user has made is over 10 but less than or equal to 40, then the customer will be labeled a “Regular customer.”
# If the maximum orders the user has made is less than or equal to 10, then the customer will be labeled a “New customer.”
ords_prods_merge['max_order'] = ords_prods_merge.groupby(['user_id'])['order_number'].transform(np.max)

# check column names for new column
ords_prods_merge.head()

# 3. Create flag

# create loyalty flag
ords_prods_merge.loc[ords_prods_merge['max_order'] > 40, 'loyalty_flag'] = 'Loyal customer'

ords_prods_merge.loc[(ords_prods_merge['max_order'] <= 40) & (ords_prods_merge['max_order'] > 10), 'loyalty_flag'] = 'Regular customer'

ords_prods_merge.loc[ords_prods_merge['max_order'] <= 10, 'loyalty_flag'] = 'New customer'

# check new column frequency
ords_prods_merge['loyalty_flag'].value_counts(dropna=False)

# check new column
ords_prods_merge[['user_id', 'loyalty_flag', 'order_number']].head(60)

# check dimensions
ords_prods_merge.shape

#export new df
ords_prods_merge.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'new_ords_prods_merged.pkl'))
