==> 8. Data Quality and consistency checks.ipynb <==
# Data Quality and consistency checks

# Contents
# 1. Investigate and understand new df
# 2. Make consistency changes
# 3. Check for mixed types, duplicates, missing values
# 4. Merge dfs

#import libraries
import pandas as pd
import numpy as np
import os

# def file path
path = r'/Users/alison/Documents/Data Analysis - Tasks/Instacart  Basket Analysis' 
#get df ords_prods_merged
ords_prods_merge = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'newest_ords_prods_merged.pkl'))

#get customer df
df_custs = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'customers.csv'), index_col=False)

# 1. Investigate and understand new df

df_custs.shape

df_custs.head()

df_custs.dtypes

# 2. Make consistency changes

# change redundant dtypes, user_id same as ords_prods_merge, change column name - Surnam to Last_name
df_custs.rename(columns = {'Surnam' : 'Last_name'}, inplace = True)

df_custs['user_id'] = df_custs['user_id'].astype('str')

df_custs['user_id'].dtype

# check changes 
df_custs

# 2. Make consistency changes

# check latest merged df
ords_prods_merge.dtypes

ords_prods_merge['user_id'] = ords_prods_merge['user_id'].astype('str')

ords_prods_merge['order_id'] = ords_prods_merge['order_id'].astype('str')

# check changes
ords_prods_merge.dtypes

# 3. Check for mixed types, duplicates, missing values

#check for mixed types
for col in df_custs.columns.tolist():
  weird = (df_custs[[col]].map(type) != df_custs[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_custs[weird]) > 0:
    print (col)

# First name has mixed types change to str
df_custs['First Name'] = df_custs['First Name'].astype('str')

#check for missing values
df_custs.isnull().sum()

# no missing values. check for duplicates
df_custs_clean = df_custs[df_custs['Last_name'].isnull() == False]
df_dups = df_custs_clean[df_custs_clean.duplicated()]

df_dups

# 4. Merge dfs

# rename merge column as stopping join on tables
ords_prods_merge.rename(columns = {'_merge' : 'old_merge'}, inplace = True)

ords_prods_merge

# merge tables on common column ensuring dtype is same
df_merged_all = df_custs.merge(ords_prods_merge, on = 'user_id', indicator = True)

df_merged_all['_merge'].value_counts()

# check dimensions
df_merged_all.shape

#export df_merged_all as pkl
df_merged_all.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'merged_all.pkl'))
