==> 10. Create new variables.ipynb <==
# 4.10 Customer analysis

# Contents
# 1. Data quality and consistency checks
# 2. Remove redundant columns
# 3. Create new variable - region
# 4. Create exclusion flag
# 5. Customer profiling

#import libraries
import pandas as pd
import numpy as np
import os

# def file path
path = r'/Users/alison/Documents/Data Analysis - Tasks/Instacart  Basket Analysis' 

# get latest merged df
all_merged = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'merged_all.pkl'))

# 1. Data quality and consistency checks

# data quality and consistency checks
all_merged.shape

all_merged.columns

# 2. Remove redundant columns

# Can remove merge and old_merge columns as not necessary for analysis. 
# Also remove first name and last name for privacy reasons. Can identify by user id. Names are not necessary for the analysis, so it is better to remove them.
all_merged = all_merged.drop(columns=['First Name', 'Last_name', 'old_merge', '_merge'])
print(all_merged)

all_merged['STATE'].value_counts()

# 3. Create new variable - region

# organised the states into four offocial regions
result = []
for value in all_merged['STATE']:
    if value in ['Maine', 'New Hampshire', 'Vermont', 'Massachusetts', 'Rhode Island', 'Connecticut', 'New York', 'Pennsylvania', 'New Jersey']:
        result.append('Northeast')
    elif value in ['Wisconsin', 'Michigan', 'Illinois', 'Indiana', 'Ohio', 'North Dakota', 'South Dakota', 'Nebraska', 'Kansas', 'Minnesota', 'Iowa', 'Missouri']:
        result.append('Midwest')
    elif value in ['Delaware', 'Maryland', 'District of Columbia', 'Virginia', 'West Virginia', 'North Carolina', 'South Carolina', 'Georgia', 'Florida', 'Kentucky', 'Tennessee', 'Mississippi', 'Alabama', 'Oklahoma', 'Texas', 'Arkansas', 'Louisiana'
]:
        result.append('South')
    else:
        result.append('West')

# create a region column for the above categories
all_merged['Region'] = result

all_merged['Region'].value_counts(dropna=False)

# cross region with spending flag to find regional spending habits
crosstab = pd.crosstab(all_merged['Region'], all_merged['spending_flag'], dropna = False)

crosstab.to_clipboard()

all_merged.shape

# export new df with Region column
all_merged.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'all_merged.pkl'))

# 4. Create exclusion flag

# create exclusion flag for customers with under 5 orders
# Loop through the order_number to set exclusion_flag
result = []

for value in all_merged['order_number']:
  if value < 5:
    result.append('Yes')
  else:
    result.append('No')

# create new column
all_merged['exclusion_flag'] = result

all_merged['exclusion_flag'].value_counts()

# Exclude rows where exclusion_flag is "Yes"
new_all_merged = all_merged[all_merged['exclusion_flag'] != 'Yes']

print(new_all_merged)

# export new df
new_all_merged.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'new_all_merged.pkl'))

# 5. Customer profiling

# get latest merged df
new_all_merged = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'new_all_merged.pkl'))

new_all_merged['Age'].describe()

# create age groups
def age_group(row):

  if row['Age'] < 26:
    return 'young adult'
  elif (row['Age'] >= 26) and (row['Age'] < 40):
    return 'adult'
  elif (row['Age'] >= 40) and (row['Age'] < 60):
    return 'middle adult'
  elif row['Age'] >= 60:
    return 'senior'
  else: return 'Not enough data'

new_all_merged['age_group'] = new_all_merged.apply(age_group, axis=1)

new_all_merged['age_group'].value_counts(dropna=False)

# create income groups
def income_group(row):

  if row['income'] < 56600:
    return 'low'
  elif (row['income'] >= 56600) and (row['income'] < 169800):
    return 'middle'
  elif row['income'] >= 169800:
    return 'high'
  else: return 'Not enough data'

new_all_merged['income_group'] = new_all_merged.apply(income_group, axis=1)

new_all_merged['income_group'].value_counts(dropna=False)

new_all_merged['department_id'].value_counts(dropna=False)

#label departments
def dept_label(row):

  if row['department_id'] == 1:
    return 'frozen'
  elif (row['department_id'] == 2):
    return 'other'
  elif (row['department_id'] == 3):
    return 'bakery'
  elif (row['department_id'] == 4):
    return 'produce' 
  elif (row['department_id'] == 5):
    return 'alcohol'
  elif (row['department_id'] == 6):
    return 'international'
  elif (row['department_id'] == 7): 
    return 'beverages'
  elif (row['department_id'] == 8):
    return	'pets'
  elif (row['department_id'] == 9):
    return	'dry goods pasta'
  elif (row['department_id'] == 10):
    return	'bulk'
  elif (row['department_id'] == 11):
    return	'personal care'
  elif (row['department_id'] == 12):
    return	'meat seafood'
  elif (row['department_id'] == 13): 
    return 'pantry'
  elif (row['department_id'] == 14): 
    return 'breakfast'
  elif (row['department_id'] == 15): 
    return 'canned goods'
  elif (row['department_id'] == 16): 
    return 'dairy eggs'
  elif (row['department_id'] == 17): 
    return 'household'
  elif (row['department_id'] == 18): 
    return 'babies'
  elif (row['department_id'] == 19): 
    return 'snacks'
  elif (row['department_id'] == 20): 
    return 'deli'
  elif (row['department_id'] == 21): 
    return 'missing'
  else: return 'Not enough data'


new_all_merged['dept_label'] = new_all_merged.apply(dept_label, axis=1)

new_all_merged['dept_label'].value_counts(dropna=False)

new_all_merged.groupby('dept_label').agg({'order_number': ['mean', 'min', 'max']})

# create dependants flag
new_all_merged.loc[new_all_merged['n_dependants'] >= 1, 'dependents_flag'] = 'Has dependents'

new_all_merged.loc[new_all_merged['n_dependants'] == 0, 'dependents_flag'] = 'No dependents'

new_all_merged['dependents_flag'].value_counts(dropna=False)

#single adult #young parent
new_all_merged.groupby('dependents_flag').agg({'Age': ['mean', 'min', 'max']})

crosstab = pd.crosstab(new_all_merged['dependents_flag'], new_all_merged['dept_label'], dropna = False)

crosstab.to_clipboard()

crosstab = pd.crosstab(new_all_merged['dependents_flag'], new_all_merged['age_group'], dropna = False)

crosstab.to_clipboard()

# create young parent profile
df1 = new_all_merged[new_all_merged['age_group']== 'young adult']
df2 = pd.crosstab(df1['dependents_flag'], df1['price_range_loc'])

mask = new_all_merged['age_group']== 'young adult'
df2 = pd.crosstab(new_all_merged.loc[mask, 'dependents_flag'], new_all_merged.loc[mask, 'price_range_loc'])

df2.head()

# create single adult profile
df1 = new_all_merged[new_all_merged['age_group']== 'young adult']
df3 = pd.crosstab(df1['fam_status'], df1['price_range_loc'])

mask = new_all_merged['age_group']== 'young adult'
df3 = pd.crosstab(new_all_merged.loc[mask, 'fam_status'], new_all_merged.loc[mask, 'price_range_loc'])

df3.head()

# export new df
new_all_merged.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'new_all_merged_plus.pkl'))
